fix(controller.py): add stop method to AcquisitionWorker to gracefully stop the acquisition loop (#3)
When attempting to exit the program, there is an issue with properly terminating the acquisition thread correctly. This can sometimes result in Python crashing and displaying an error message. 

To address this problem, I have added a stop method to ﻿acquisition_worker that allows for a graceful termination of the process, and with `self.acquisition_thread.quit()` and ﻿`self.acquisition_thread.wait()` to ensure that the thread is safely terminated.